LayoutTests/accessibility/ios-simulator/image-overlay-elements-expected.txt

PASS axImageOverlayElements.length is 2
hello
world
PASS successfullyParsed is true

TEST COMPLETE
